refactor utils to multiple ts
